@@ -1,5 +1,5 @@
 #!/usr/bin/env python
-# -*- coding: utf-8 -*-
+"""Synology DSM setup."""
 
 # NOTE(ProtoThis) Guidelines for Major.Minor.Micro
 # - Major means an API contract change
@@ -7,7 +7,6 @@
 # - Micro means change of any kind (unless significant enough for a minor/major).
 
 from setuptools import setup, find_packages
-from codecs import open
 
 REPO_URL = "https://github.com/ProtoThis/python-synology"
 VERSION = "0.9.0"
@@ -25,23 +24,24 @@ setup(
     download_url=REPO_URL + "/tarball/" + VERSION,
     description="Python API for communication with Synology DSM",
     long_description=long_description,
-    author="FG van Zeelst (ProtoThis)",
+    author="Quentin POLLET (Quentame) & FG van Zeelst (ProtoThis)",
     packages=find_packages(include=["synology_dsm*"]),
     install_requires=required,
-    python_requires=">=2.7, !=3.0.*, !=3.1.*, !=3.2.*, !=3.3.*",
+    python_requires=">=3.6",
     license="MIT",
     classifiers=[
+        "Development Status :: 5 - Production/Stable",
         "Intended Audience :: Developers",
         "License :: OSI Approved :: MIT License",
         "Operating System :: OS Independent",
         "Programming Language :: Python",
-        "Programming Language :: Python :: 2.7",
+        "Programming Language :: Python :: 3 :: Only",
         "Programming Language :: Python :: 3",
-        "Programming Language :: Python :: 3.4",
-        "Programming Language :: Python :: 3.5",
         "Programming Language :: Python :: 3.6",
         "Programming Language :: Python :: 3.7",
         "Programming Language :: Python :: 3.8",
+        "Programming Language :: Python :: 3.9",
+        "Topic :: Software Development :: Libraries",
     ],
     keywords=["synology-dsm", "synology"],
 )
