@@ -1,4 +1,3 @@
-# -*- coding: utf-8 -*-
 """Library exceptions."""
 from .const import API_AUTH, ERROR_AUTH, ERROR_COMMON, ERROR_DOWNLOAD_SEARCH, ERROR_DOWNLOAD_TASK, ERROR_FILE, ERROR_SURVEILLANCE, ERROR_VIRTUALIZATION
 
@@ -24,7 +23,7 @@ class SynologyDSMException(Exception):
             reason = "Unknown"
         
         error_message={"api": api, "code": code, "reason": reason, "details": details}
-        super(SynologyDSMException, self).__init__(error_message)
+        super().__init__(error_message)
 
 # Request
 class SynologyDSMRequestException(SynologyDSMException):
@@ -34,57 +33,56 @@ class SynologyDSMRequestException(SynologyDSMException):
         ex_reason = exception.args[0]
         if hasattr(exception.args[0], "reason"):
             ex_reason = exception.args[0].reason
-        message = "%s = %s" % (ex_class, ex_reason)
-        super(SynologyDSMRequestException, self).__init__(None, -1, message)
+        super().__init__(None, -1, f"{ex_class} = {ex_reason}")
 
 # API
 class SynologyDSMAPINotExistsException(SynologyDSMException):
     """API not exists exception."""
     def __init__(self, api):
-        super(SynologyDSMAPINotExistsException, self).__init__(api, -2, "API %s does not exists" % api)
+        super().__init__(api, -2, f"API {api} does not exists")
 
 class SynologyDSMAPIErrorException(SynologyDSMException):
     """API returns an error exception."""
     def __init__(self, api, code, details):
-        super(SynologyDSMAPIErrorException, self).__init__(api, code, details)
+        super().__init__(api, code, details)
 
 # Login
 class SynologyDSMLoginFailedException(SynologyDSMException):
     """Failed to login exception."""
     def __init__(self, code, details=None):
-        super(SynologyDSMLoginFailedException, self).__init__(API_AUTH, code, details)
+        super().__init__(API_AUTH, code, details)
 
 
 class SynologyDSMLoginInvalidException(SynologyDSMLoginFailedException):
     """Invalid password & not admin account exception."""
     def __init__(self, username):
-        message = "Invalid password or not admin account: %s" % username
-        super(SynologyDSMLoginInvalidException, self).__init__(400, message)
+        message = f"Invalid password or not admin account: {username}"
+        super().__init__(400, message)
 
 
 class SynologyDSMLoginDisabledAccountException(SynologyDSMLoginFailedException):
     """Guest & disabled account exception."""
     def __init__(self, username):
-        message = "Guest or disabled account: %s" % username
-        super(SynologyDSMLoginDisabledAccountException, self).__init__(401, message)
+        message = f"Guest or disabled account: {username}"
+        super().__init__(401, message)
 
 
 class SynologyDSMLoginPermissionDeniedException(SynologyDSMLoginFailedException):
     """No access to login exception."""
     def __init__(self, username):
-        message = "Permission denied for account: %s" % username
-        super(SynologyDSMLoginPermissionDeniedException, self).__init__(402, message)
+        message = f"Permission denied for account: {username}"
+        super().__init__(402, message)
 
 
 class SynologyDSMLogin2SARequiredException(SynologyDSMLoginFailedException):
     """2SA required to login exception."""
     def __init__(self, username):
-        message = "Two-step authentication required for account: %s" % username
-        super(SynologyDSMLogin2SARequiredException, self).__init__(403, message)
+        message = f"Two-step authentication required for account: {username}"
+        super().__init__(403, message)
 
 
 class SynologyDSMLogin2SAFailedException(SynologyDSMLoginFailedException):
     """2SA code failed exception."""
     def __init__(self):
         message = "Two-step authentication failed, retry with a new pass code"
-        super(SynologyDSMLogin2SAFailedException, self).__init__(404, message)
+        super().__init__(404, message)
